# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ninotek F&B POS API
  version: 1.0.0
  description: |
    Hợp đồng REST cho toàn hệ thống Ninotek F&B POS.

    **Đây là nguồn sự thật duy nhất cho API.** Client C# (NinoPOS, qua NSwag) và
    Dart (NinoOrder/NinoDash, qua openapi-generator) đều sinh code từ file này.
    CI sẽ fail nếu implementation lệch khỏi contract.

    ## Hai mặt phẳng API tách biệt

    | | LAN API | Cloud API |
    |---|---|---|
    | Máy chủ | NinoPOS tự host (Kestrel) | NinoCloud (NestJS) |
    | Địa chỉ | `http://<ip-máy-POS>:8080` | `https://api.ninotek.vn` |
    | Client | NinoOrder (tablet nhân viên) | NinoPOS (sync), NinoDash |
    | Xác thực | Device token cấp khi ghép nối | JWT Bearer |
    | Yêu cầu Internet | **Không** — chạy hoàn toàn nội bộ | Có |

    Khi mất Internet, LAN API vẫn phục vụ bình thường. Đây là nền tảng của
    kiến trúc Offline-First.

    ## Optimistic Locking

    Mọi endpoint sửa dữ liệu đều nhận `rowVersion` mà client đọc được lúc đầu.
    Nếu bản ghi đã bị người khác sửa, server trả **409 Conflict** kèm bản ghi
    hiện tại để client hiển thị và cho người dùng quyết định.

  contact:
    name: Ninotek — Công nghệ & Giải pháp số
    url: https://ninotek.vn
  license:
    name: Proprietary
    url: https://ninotek.vn/license
servers:
  - url: http://{posHost}:8080
    description: LAN API — NinoPOS tự host trong mạng nội bộ quán
    variables:
      posHost:
        default: 192.168.1.10
...
  - url: https://api.ninotek.vn
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/cloud/payments/vietqr/generate:
    post:
      tags: [Cloud — Thanh toán]
      summary: Sinh mã VietQR động cho hoá đơn
      description: |
        Chuỗi trả về tuân thủ EMVCo + NAPAS VietQR/IBFT.

        `addInfo` **phải** là `orders.order_code` (chỉ gồm `[A-Z0-9]`) vì đó là
        khoá duy nhất để đối soát khi webhook ngân hàng báo về. Chuỗi có dấu
        tiếng Việt hoặc ký tự đặc biệt sẽ bị chuẩn hoá; nếu mã đơn không sống
        sót qua bước chuẩn hoá thì không thể đối soát tự động.

        Bộ golden test vectors:
        `packages/api-contracts/data/vietqr-test-vectors.json`
      operationId: cloudGenerateVietQr
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VietQrGenerateRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VietQrGenerateResponse' }
        '422':
          description: Dữ liệu không hợp lệ (acqId sai, số tiền <= 0, ...)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/cloud/payments/vietqr/webhook:
    post:
      tags: [Cloud — Thanh toán]
      summary: Nhận IPN báo chuyển khoản thành công từ ngân hàng / cổng thanh toán
      description: |
        Gọi bởi Casso / SePay / cổng ngân hàng — **không phải** bởi app của Ninotek.

        Ba quy tắc bắt buộc khi hiện thực:

        1. **Xác thực HMAC-SHA256 trước khi đọc body.** Endpoint này public trên
           Internet; không xác thực chữ ký nghĩa là ai cũng có thể giả lệnh
           "đã thanh toán" và ăn hàng miễn phí.
        2. **Đối soát mã đơn bằng regex, không so sánh `==`.** Ngân hàng chèn
           thêm chữ vào nội dung chuyển khoản, ví dụ
           `"MBVCB.9876543210.NINOPOS100234.CT tu 0123456789"`.
        3. **Idempotent theo `transactionId`.** Ngân hàng bắn lại IPN khi chưa
           nhận được 200. Ghi trùng = ghi nhận doanh thu hai lần.

        Luôn trả 200 kể cả khi không khớp đơn nào, để cổng thanh toán không
        retry vô hạn — ghi vào `payment_webhook_logs` rồi xử lý thủ công sau.
      security:
        - WebhookSignature: []
      operationId: cloudVietQrWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VietQrWebhookPayload' }
      responses:
        '200':
          description: Đã tiếp nhận (kể cả khi không khớp đơn nào)
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { type: boolean }
                  matchedOrderCode: { type: [string, 'null'] }
        '401':
          description: Chữ ký HMAC không hợp lệ
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT ngắn hạn (15 phút) cho Cloud API; device token cho LAN API.
    WebhookSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 của raw body với secret chia sẻ với cổng thanh toán.
...
  schemas:
    Problem:
      type: object
      description: Lỗi theo chuẩn RFC 9457 (Problem Details for HTTP APIs).
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, example: 'https://ninotek.vn/errors/table-locked' }
        title: { type: string, example: Bàn đang bị nhân viên khác thao tác }
        status: { type: integer, example: 409 }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
...
    VietQrGenerateRequest:
      type: object
      required: [orderId, accountNo, accountName, acqId, amount, addInfo]
      properties:
        orderId: { type: string, format: uuid }
        accountNo: { type: string, pattern: '^[A-Za-z0-9]{1,19}$', example: '0123456789' }
        accountName: { type: string, example: CONG TY TNHH CONG NGHE & DICH VU NINOTEK }
        acqId:
          type: string
          pattern: '^\d{6}$'
          example: '970436'
          description: Mã BIN ngân hàng. Tra trong packages/api-contracts/data/napas-bank-bins.json
        amount: { type: integer, minimum: 1, maximum: 9999999999, example: 150000 }
        addInfo:
          type: string
          maxLength: 95
          example: NINOPOS100234
          description: |
            PHẢI là `orders.order_code`. Tối đa 95 ký tự (tag 62 của EMVCo giới hạn 99,
            trừ 4 ký tự header của tag 08). Nhiều ngân hàng còn cắt quanh mốc 50 ký tự
            khi bắn webhook, nên mã đơn phải nằm ở ĐẦU chuỗi.
        template: { type: string, enum: [compact, qr_only, print], default: compact }
    VietQrGenerateResponse:
      type: object
      required: [qrCode, qrDataURL, expiresAt]
      properties:
        qrSessionId: { type: string, format: uuid }
        qrCode:
          type: string
          description: Chuỗi EMVCo thô, đã kèm CRC-16/CCITT-FALSE
          example: 00020101021238540010A00000072701240006970436011001234567890208QRIBFTTA530370454061500005802VN62170813NINOPOS10023463048507
        qrDataURL: { type: string, description: 'PNG dạng data:image/png;base64,...' }
        expiresAt: { type: string, format: date-time }
        websocketUrl:
          type: string
          example: ws://localhost:8080/ws/payments?order_id=NINOPOS100234
          description: NinoPOS mở kênh này để nhận PAYMENT_SUCCESS và tự chốt bill.
    VietQrWebhookPayload:
      type: object
      required: [gateway, transactionId, amount, content, status]
      properties:
        gateway: { type: string, example: SePay }
        transactionId: { type: string, example: FT2624490182, description: Khoá idempotency }
        amount: { type: number, example: 150000 }
        content:
          type: string
          example: MBVCB.9876543210.NINOPOS100234.CT tu 0123456789
          description: Nội dung chuyển khoản THÔ. Ngân hàng chèn thêm chữ — phải dò mã đơn bằng regex.
        orderCode: { type: [string, 'null'], description: Một số cổng đã tách sẵn; không phải lúc nào cũng có }
        bankAccount: { type: string }
        status: { type: string, enum: [SUCCESS, FAILED] }
        timestamp: { type: string, format: date-time }
...